@@ -1,6 +1,6 @@
 {
   "name": "react-dom",
-  "version": "17.0.2",
+  "version": "18.0.0-beta-24dd07bd2-20211208",
   "description": "React package for working with the DOM.",
   "main": "index.js",
   "repository": {
@@ -19,10 +19,10 @@
   "dependencies": {
     "loose-envify": "^1.1.0",
     "object-assign": "^4.1.1",
-    "scheduler": "^0.20.2"
+    "scheduler": "0.21.0-beta-24dd07bd2-20211208"
   },
   "peerDependencies": {
-    "react": "17.0.2"
+    "react": "18.0.0-beta-24dd07bd2-20211208"
   },
   "files": [
     "LICENSE",
@@ -34,12 +34,12 @@
     "server.browser.js",
     "server.node.js",
     "test-utils.js",
+    "unstable_testing.js",
     "cjs/",
     "umd/"
   ],
   "browser": {
-    "./server.js": "./server.browser.js",
-    "./unstable-fizz.js": "./unstable-fizz.browser.js"
+    "./server.js": "./server.browser.js"
   },
   "browserify": {
     "transform": [
