@@ -1,14 +1,13 @@
-import { StrictMode } from "react";
-import ReactDOM from "react-dom";
+import * as ReactDOM from "react-dom";
 import { Root } from "./features";
 import reportWebVitals from "./reportWebVitals";
 
-ReactDOM.render(
-  <StrictMode>
-    <Root />
-  </StrictMode>,
-  document.getElementById("root")
-);
+const container = document.getElementById("root");
+
+// @ts-expect-error `ReactDOM.createRoot` is not yet included on "@types/react".
+const root = ReactDOM.createRoot(container);
+
+root.render(<Root />);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
