@@ -8,7 +8,7 @@ import foodEntryRoutes from './routes/foodEntry.routes';
 const app = express();
 
 const corsOptions = {
-  origin: 'http://localhost:3000',
+  origin: '*',
 };
 
 app.use(cors(corsOptions));
